добавили функцию покупки акций

+++ src/main/java/com/tg/investbot/helper/CommandHelper.java
@@ -2,6 +2,9 @@ package com.tg.investbot.helper;
 
 import com.tg.investbot.registry.UserCommandName;
 
+import java.lang.reflect.Array;
+import java.util.ArrayList;
+
 public class CommandHelper {
     public static UserCommandName getCommandName(String message) {
         var parameters = message.split(" ");
@@ -12,7 +15,7 @@ public class CommandHelper {
         return message.split(" ")[1];
     }
 
-    public static Long getCount(String message) {
-        return Long.parseLong(message.split(" ")[2]);
+    public static Integer getCount(String message) {
+        return Integer.parseInt(message.split(" ")[2]);
     }
 }
